requirements.txt
streamlit
pillow 
pandas 
qrcode 
tqdm
